+++ nextjs/src/components/layout/header.tsx
@@ -6,15 +6,15 @@ import { setOrgId, setOrgInfo } from '@/lib/orgSlice'
 import { useAppDispatch, useAppSelector } from '@/lib/hooks'
 
 import { Breadcrumbs } from '../breadcrumbs'
+import { OrgSwitcher } from '../org-switcher'
 import { Organisation } from '@/features/dashboard/type/organization'
-import { Separator } from '../ui/separator'
 import { SidebarTrigger } from '../ui/sidebar'
 import { UserNav } from './user-nav'
 import { getOrganizations } from '@/app/api/organization'
 
 export default function Header(): React.JSX.Element {
   const dispatch = useAppDispatch()
-  const [, setOrgList] = useState<Organisation[]>([])
+  const [orgList, setOrgList] = useState<Organisation[]>([])
   const tenantId = useAppSelector((state) => state.organization.orgId)
 
   useEffect(() => {
@@ -58,35 +58,30 @@ export default function Header(): React.JSX.Element {
     fetchOrganizations()
   }, [])
 
-  // const handleSwitchTenant = (orgId: string): void => {
-  //   const selected = orgList.find((org) => org.id === orgId)
-  //   if (selected) {
-  //     dispatch(setOrgId(selected.id))
-  //     dispatch(
-  //       setOrgInfo({
-  //         id: selected.id,
-  //         name: selected.name,
-  //         description: selected.description,
-  //         logoUrl: selected.logoUrl,
-  //         roles:
-  //           selected.userOrgRoles?.map(
-  //             (role: { orgRole: { name: string } }) => role?.orgRole?.name,
-  //           ) || [],
-  //       }),
-  //     )
-  //   }
-  // }
+  const handleSwitchTenant = (orgId: string): void => {
+    const selected = orgList.find((org) => org.id === orgId)
+    if (selected) {
+      dispatch(setOrgId(selected.id))
+      dispatch(
+        setOrgInfo({
+          id: selected.id,
+          name: selected.name,
+          description: selected.description,
+          logoUrl: selected.logoUrl,
+          roles:
+            selected.userOrgRoles?.map(
+              (role: { orgRole: { name: string } }) => role?.orgRole?.name,
+            ) || [],
+        }),
+      )
+    }
+  }
 
   return (
     <header className="flex h-16 shrink-0 items-center justify-between gap-2 transition-[width,height] ease-linear group-has-data-[collapsible=icon]/sidebar-wrapper:h-12">
       <div className="flex items-center gap-2 px-4">
         <SidebarTrigger className="-ml-1" />
-        <Separator orientation="vertical" className="mr-2 h-4" />
-        <Breadcrumbs />
-      </div>
-
-      <div className="flex items-center gap-2 px-4">
-        {/* <OrgSwitcher
+        <OrgSwitcher
           tenants={orgList.map((org) => ({
             id: org.id,
             name: org.name,
@@ -94,7 +89,11 @@ export default function Header(): React.JSX.Element {
           }))}
           defaultTenant={orgList.length > 0 ? orgList[0] : undefined}
           onTenantSwitch={handleSwitchTenant}
-        /> */}
+        />
+        <Breadcrumbs />
+      </div>
+
+      <div className="flex items-center gap-2 px-4">
         {/* NOTE: Currently disabling search and mode toggle */}
         <div className="hidden md:flex">{/* <SearchInput /> */}</div>
         {/* <ModeToggle /> */}
